Fix broken licence agreement tear-down (#673)
https://eaflood.atlassian.net/browse/WATER-4336

In [Improve tear down speed](https://github.com/DEFRA/water-abstraction-system/pull/671) we made sweeping changes to the tear-down service which is used to support our [water-abstraction-acceptance-tests](https://github.com/DEFRA/water-abstraction-acceptance-tests). The new changes improved the performance of the clean-down.

However, we've spotted that one of the delete statements got lost in the whirlwind of our strategy testing; licence agreements not marked as 'test' but linked to 'test' licences. This a common issue the tear-down service has to deal with due to the way the tests are currently written and the inherited test seed process works.

This change adds the missing statement.