Switch to single query strategy
We started with the original aim of calling the triggers before and after the main delete statements for each schema. But then we also spotted that there are occasions where we also should have been controlling the order of the deletions, the billing tables being a primary example.

To determine which billing transactions to delete you have to be able to link from them to the 'test' region via billing invoice licences, billing invoices, and then billing batches. If we delete the test data from one of these linking tables before the delete billing transactions statement gets called nothing will get cleansed. This might be the real reason why the **Promise all** strategy came out tops; it wasn't always deleting everything.

We've also taken this opportunity to split the crm service into `crm` and `crm_v2` so we have a one-to-one mapping of tear-down service and schema.

Finally, we've updated the README to record the change in strategy and reason why.