Fix licence agreement tear down again!
https://eaflood.atlassian.net/browse/WATER-4336

Ok, we've made a bit of a hash with trying to [improve the tear-down speed](https://github.com/DEFRA/water-abstraction-system/pull/671).

We realised in the re-write we'd dropped a licence agreement DELETE statement that needed including so [fixed it](https://github.com/DEFRA/water-abstraction-system/pull/673/files).

Then we realised the tear-down was now faster but more flaky so [switched to the single query strategy](https://github.com/DEFRA/water-abstraction-system/pull/674).

Only in the switch what do you know, we lost the licence agreement delete statement again.

Hopefully for the last time, we fix this!