Update 2PT /check endpoint with latest services
https://eaflood.atlassian.net/browse/WATER-4339

This follows on from [Update 2PT check endpoint with alternate changes](https://github.com/DEFRA/water-abstraction-system/pull/582) where we applied the services we refactored out of our 2PT spike branch to the `/check` endpoint.

The final set of services have now been refactored and added to `main`, so we can complete the work of replacing the stand-ins currently being used.