Add tests to the licence model
We want the custom instance method tested within the model tests. Because these instance methods rely on the right data being available in the instance we also end up testing the modifier. This was the approach we took on [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api); we add modifiers because numerous things depend on them and they are better tested through those things.