Update stand-ins service
We can drop the PrepareLicences service and use the real one. However, when verifying the results we find it useful to see which lines were allocated to which elements. This is not something we care about when generating the review results for real.

The problem is these can only happen deep inside the allocation process. We don't want to add toggles to the production service just for this. We're on the cusp of having the pages ready to go at which point we expect the `/check` endpoint to become defunct.

So, for now we just copy the real service into our stand-ins and add the tweak which captures which lines were allocated.