Create licence holder seeder (#689)
https://eaflood.atlassian.net/browse/WATER-4188

Whilst carrying out testing there have been multiple instances where we need to create data for a licence holder. This isn't a simple process and involves creating records in 4 tables in the DB.

We have therefore created a "seeder" which when passed a licence reference will create the records required for a licence holder.