Thoughts and opinions - @cruikshanks contributions
Having deep dived on what we have so far we've in the main managed to solve the problem of validating a date both generically (is this date valid) and from a business perspective (is this date acceptable).

There is an issue of handling no selection being made (it looks like this got broken with later changes). But IMHO I think there are also some simplifications and optimisations that could be made.

In the subsequent changes I'll try and make the change and explain why I'm doing it. I'll deal with all the tests I break at the end!