Move submit logic to dedicated service
This deals with 2 problems I foresaw.

Optimisation

The first was the current logic in the controller depended on calling the `StartDateService` twice in the event of a validation error. From what I could see this was purely to access the presenter to generate the page data with the error and payload included. But it meant a second query request was going out to get the 'session', which we already had.

Simplification

Our convention of testing controller's via 'controller' tests (integration in other words) is intended. The more complex the controller, the more 'controller' tests we'll have to add that will be slower and more complex to setup. To get fewer, easier to write 'controller' tests we need to make to our controllers as 'dumb' as possible.

This is why we follow a pattern of doing all the work in a 'service' and I felt this controller would benefit from one. But rather than reuse the `StartDateService` if we created a `SubmitStartDateService` it would allow us to also handle our optimisation issue.

---

We've not made any changes to how the thing is validated, the formatting of the data or the view template in this change. We've just moved what was in the controller to a new service.