Consistency - update quotes in view template
To stop anyone second guessing about what is the 'right' way to do something we try and focus on being as consistent as possible. We haven't been, and views being a recent addition to the project is where we're most at fault.

But trying to be an exemplar this updates the quotes usage in the template to match what StandardJS would shout at us about in our real code; single quotes as the default, doubles when we need them.

We also fix some inconsistent indentation.