Separation & Simplification - move items to view
Separation

Again, we have the excuse that we are still working out our conventions for where things sit when it comes to building pages. We know we want as little logic in the views as possible. Nothing beyond some looping, string concatenation and if/else should be in the view.

But what is emerging is that anything to do with the structure of the page should sit in the view. For example, we have a page with a table. We could have the presenter combine both the layout, data and element options in one go. But that mixes concerns; the presenter is now responsible for both the data and the layout.

So, if we can we should leave layout and structure duties to the view and data handling to the code/presenters.

Simplification

One benefit of this is it has made our presenter much simpler. There is some increase in complexity in the view. But we think this is nullified by most of the changes just being structural. We have only added one new conditional to the view, but removed a great many from the presenter.

Admittedly, some of this simplification comes from the fact we are no longer trying to style individual elements. But we still think separating the concerns has made things clearer in both the view and the code.