Simplification - use static value for start date
We don't need the licence version start date value to be passed into the form and then returned to us. Whether it is the GET or the POST we'll have this information in the session object.

So, we can simplify and be consistent by just using a static, set value.