Simplify - remove formatting of start and end date
Now we are not passing the date through to the view for use in the form. So, that is one less reason to format it. Secondly, the Joi date validators accept a string. So, what is in the session can be passed straight through.

So, rather than having the presenter format the dates, we can pass them straight through to the validator from the session object. It also removes the condition of checking for an error and calling the presenter for a second time. Now we can just call it once.