Convention & consistency - use existing function
Mainly to be consistent we replace the direct use of `pad()` with our own helper function. We also rename the function. Our convention generally sways to a 'ruby' style.

We actively avoid prefixing anything with `get` and `set`. But also where a function is returning a named value, we err on the side of naming the function after the named value. The fact we're calling a function is an indication we're doing something more than popping it off an object. But we make our lives a little easier by not stressing what to name that 'something'!