Simplification - remove date field validation
Being able to highlight a specific field being incorrect I believe is a 'nice to have'. To do so is going to add lots more complexity to the code and open a series of other scenarios we have to cater for. Which do we highlight if someone enters 'Z 13 [null]'? Do we highlight all 3 or just the first one?

Add to that the guidance in the design system is clear that if highlighting specific fields the error message needs to include this information.

> If you're highlighting just one field - either the day, month or year - only style the field that has an error. The error message must say which field has an error, [..]

The ACs for the story only require us to check the date is valid and meets certain business rules. Where the date is invalid we have been given only one message to display, 'Enter a real start date'.

So, as we are not providing distinct error messages and we want to make maintenance of the service as simple as possible we remove all the functionality related to checking individual date fields.