Fix - getting Joi date format error
With the last change we made we are now passing in an invalid date format to Joi to validate. Because it wasn't covered when we enter a partial date or one including invalid values we see `"fullDate" must be in ISO 8601 date format`.

Adding this message gets it displaying `Enter a real start date` again.