Documentation - add JSDoc for the validator
As a minimum we include a JSDoc comment for all our public methods. This can be a great help when using a tool like VSCode as it will display the comment in the ID when you hover over references to it elsewhere in the code. It also allows us to give some context to why the function exists.

We have more guidance about this in https://github.com/DEFRA/water-abstraction-team/blob/main/coding_conventions.md#jsdoc-comments