Fix - select anotherStartDate on nothing selected
This change fixes one of the original issues found when reviewing the change. If no option was selected then the page would reload with 'Another date' automatically selected.

By including the selected options as properties in our page data we can more reliably tell the input controls which radio button to select, or none at all.

We did have to go back on a previous decision though. We wanted to keep the schema as a single thing. However, we found we could not get the different error states to fire properly. Splitting the schema and only validating against each got the screen behaving as expected.