Fix - ensure form is highlighted on nothing
In previous changes we solved the issue of the 'Another date' radio button being selected when nothing was selected and just a Joi error message being shown.

The final piece of the puzzle is to ensure that the whole control gets highlighted when no option is selected which this change does.