Fix - using wrong start date for validation
The acceptance criteria in the ticket is

> I enter a date that is earlier than the original licence effective (start date) date

The original licence effective date is the start date recorded against the licence and not what is against the current licence version record. When you search for a licence it is the **Effective from** date you see on the Summary tab.

We were not capturing that in the session. So, this change updates the `InitiateReturnRequirementService` to fetch and set it against the session as `startDate`. We move the licence version start date to `currentVersionStartDate` and update the presenter and the view.