Fix - broken journey
We'd overwritten the logic that was in the controller used to determine which page to go to next. Start date is the first page you go to for both journeys. But once submitted we have to redirect to different pages depending on whether the user clicked 'Set up new returns requirements' or 'No returns required'.