Fix - date validation using joi-date extension
When fixing the tests for the `StartDateValidator` we wrote one to confirm that a this payload would cause a validation error.

```javascript
{
  'start-date-options': 'anotherStartDate',
  'start-date-day': '29',
  'start-date-month': '2',
  'start-date-year': '2023'
}
```

There was no leap year in 2023 so this is not a real date. But the test failed; Joi was letting it through. When we checked the value it was generating was '2023-03-01'. This then triggered a memory of an issue we'd had with [date validation on the sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api/pull/170).

That was more about Joi mixing up US and UK date formats but it reminded us about this comment in the Joi docs

> If the validation convert option is on (enabled by default), a string or number will be converted to a Date if specified. Note that some invalid date strings will be accepted if they can be adjusted to valid dates (e.g. '2/31/2019' will be converted to '3/3/2019') by the internal JS Date.parse() implementation.
>
> https://joi.dev/api/?v=17.12.0#date

This explains exactly what we were seeing in the test. JavaScript's [Date.parse()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/parse) is 'helpfully' converting it to a valid date.

We clearly don't want this to happen. So, we've come to the same conclusion. The only way to prevent this is to add the [joi-date extension](https://joi.dev/module/joi-date/) and be explicit about the format we are validating.