https://eaflood.atlassian.net/browse/WATER-4324
> This is part of a series of changes to allow us to replace the legacy view licence page

When viewing a licence using the [legacy water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui) the page will display information on who the registered user is.

The registered is an external user who has declared that they manage the licence via the external UI. They are not the licence holder, just the person declared to be responsible for managing it.

Unfortunately, this functionality dates from the very start of the service and the data is in one of the oldest legacy schemas; `crm`. It seems this was one of the things which the previous team never managed to migrate to `crm_v2`, their intended better CRM structure.

We are expected to show this information in the licence view so we need to create some new models. The first of these was  [LicenceEntityModel](https://github.com/DEFRA/water-abstraction-system/pull/690), which is based on `crm.entity` and just holds a name and type.

The second is `LicenceEntityRoleModel`, which is based on `crm.entity_roles`. It appears to just be a joining table between `crm.entity` and `crm.roles`. But we need it because it identifies which `crm.entity` has the role of 'primary_user'.