Correct LicenceEntity changes
In creating `LicenceEntityRole` we spotted that the ID is not automatically generated. When we checked back on `LegacyEntity` we found the same.

We also spotted that though we can see UUID's have been used for both tables ID columns the type is just 'string'.

So, this change corrects what we've just done in [Add LicenceEntityModel](https://github.com/DEFRA/water-abstraction-system/pull/690).