DO NOT MERGE - licence holder changes
These are some of the changes made in https://github.com/DEFRA/water-abstraction-system/pull/679 to support displaying the licence holder in the view page. We will not retain them at the end of this PR. But `FetchLicenceService` is where we intend to first make use of our new registered user modifier. To make sure it works and doesn't break anything we're copying what will soon be merged before we then start playing around!