Alter LicenceDocumentHeader view migration
There were fields which didn't appear to be used when we create the migration. No doubt we didn't check pre-prod properly because if we had we would have seen that a number of these fields are populated.

Specifically, we need `companyEntityId` in order to link through to the new tables we've added to get the primary user. And we need `document_name` as this is where **Licence name** comes from (shown as a field in the external summary and above the licence ref in the internal summary).