Get primary user name and licence name
Here is a working example of the query we need to add to get a licence's registered username and if set the licence name they assigned.

We're having to use an standard join rather than a `withGraphFetched` because the link between `LicenceDocumentHeader` and `LicenceEntityRole` is via the `companyEntityId` which just seems _too_ tenuous.

So, we haven't defined a relationship for it in the model hence the need for the stand INNER JOIN.

We're also having to drop down into Knex for part of it in order to add an extra clause to the JOIN. We only want the record that is set as the `primary_user`. Because we are using standard joins we can't deal with this in the WHERE clause. We could have left it out but then it'd be prudent to add extra logic to ensure the `LicenceEntityRole` definitely has the role `primary_user`. This means you can use the result without extra checking.