Move the calculate & log time feature to lib
https://eaflood.atlassian.net/browse/WATER-4345

We're about to start working on rebuilding the SROC annual bill run engine in this project. To do that we'll be referencing the existing supplementary billing code a lot.

One of the first things we want to bring across is the ability to record how long it takes to generate the bill run. You'll see this in a number of places in our project; a function will initialise a variable called `startTime` and when the work is complete pass it to a function called `_calculateAndLogTime()` which will output a message that includes the time taken.

In fact you'll see this in so many places that it is time to say 'enough is enough!'

This change moves that functionality to the `lib/` and then updates the project to reuse it rather than duplicating `_calculateAndLogTime()` all over the place.