Handle create annual bill run requests (#697)
https://eaflood.atlassian.net/browse/WATER-4347

The system already handles create bill run requests. But only for supplementary and two-part tariff bill runs. We need to amend our logic to also handle requests to create annual bill runs.

There is still plenty of work to do processing and generating the bill run. But this change will at least get the bill run record created and set to a state of `QUEUED` if only so we can confirm our [changes to the legacy UI to trigger it](https://github.com/DEFRA/water-abstraction-ui/pull/2503) are working.