Fix licences is water undertaker column
In our test legacy table we created the water.licences.is_water_undertaker as a boolean and correctly didn't set a default. But just to make things 'fun' the previous team also made the column not nullable.

We've overlooked that meaning the tests have been getting away with not setting a value. This change ensures the helper always sets a value but also updates the test legacy table to match the real thing.

NOTE: Take a look at the migration and what we have had to do. This is going to apply to all migrations that try to alter a legacy table so worth knowing about.