Fix test legacy charge elements table
The charge references table in new money!

We found a new test was failing because the volume between what was found in the record didn't match what we thought we were setting it volume to in the helper.

When we checked it was because the legacy table has the column as an unconstrained `NUMERIC` (it can hold any value up to implementation limits). When we use the Knex `decimal()` type it creates a `NUMERIC(8,2)` by default.

> For reference you can create an unconstrained numeric using `decimal('my_column', null, null)`.

So, we we thought we were saving `6.819` to the DB but in fact it was being saved as `6.82`. This migration script corrects the legacy test table.