Update view to reference error
We adding an error property to the page data to be consistent across the pages. As we learnt on the 'Select a start date page' we might need something more complex than just the text message. Having an `error:` object gives some consistency to the page data and the flexibility to add whatever need (or not).