Thoughts and opinions
Having completed the [Returns required journey - Select start date page iteration 2 (1 of 7)](https://github.com/DEFRA/water-abstraction-system/pull/646) I was expecting the pattern implemented to follow that more than what we'd originally done for 'no-returns-required'.

But with so few pages implemented with controls it's hard to see a pattern! So, that wasn't made clear enough. To help I've completed [Update no-returns-required to use submit service](https://github.com/DEFRA/water-abstraction-system/pull/708) which updates the 'no-returns-required' page to follow the template laid out for 'start-date'. Essentially, validation and error handling will now be done in a `Submit[Page]Service` rather than reuse `[Page]Service`.

And so again, I'll go through and contribute to this PR to highlight the changes and when possible, expand on the reasons why.