Consistency - add new submit service
This is slightly pre-emptive but we know we will need to handle persisting valid payloads to the sessions record soon.

Having a service separate from what we call on the GET request gives us an appropriate place to put that. It also allows us to simplify the controller logic in the meantime.