Add controls & validation to rtn reqs reason page (#688)
* Returns required journey - Select return reason page iteration 2 (2 of 7)

https://eaflood.atlassian.net/browse/WATER-4265

This is the second page in the returns required journey.

This page will have radio control options.

This page will have validation to ensure that the user has selected a radio option.

* feat(app): WATER-4265 Added controls and validation for step 2 of 7

This commit adds controls for the 2nd step in the no returns required journey. It also has validation which checks to ensure a value is selected on the form

* fix(app): WATER-4265 fixed broken test

test had wrong data inside. fixed with correct data

* fix: fix broken test

* fix: amend sonarcloud issues

* fix: amend another sonarcloud issue

* Thoughts and opinions

Having completed the [Returns required journey - Select start date page iteration 2 (1 of 7)](https://github.com/DEFRA/water-abstraction-system/pull/646) I was expecting the pattern implemented to follow that more than what we'd originally done for 'no-returns-required'.

But with so few pages implemented with controls it's hard to see a pattern! So, that wasn't made clear enough. To help I've completed [Update no-returns-required to use submit service](https://github.com/DEFRA/water-abstraction-system/pull/708) which updates the 'no-returns-required' page to follow the template laid out for 'start-date'. Essentially, validation and error handling will now be done in a `Submit[Page]Service` rather than reuse `[Page]Service`.

And so again, I'll go through and contribute to this PR to highlight the changes and when possible, expand on the reasons why.

* Housekeeping - clean up after merging in `main`

* Correction - The module name does not match file

* Correction - wrong page in comments

That was me! Doh!

* Consistency - add new submit service

This is slightly pre-emptive but we know we will need to handle persisting valid payloads to the sessions record soon.

Having a service separate from what we call on the GET request gives us an appropriate place to put that. It also allows us to simplify the controller logic in the meantime.

* Remove error handling from the presenter

Now handled by the submit service.

* Remove optional error from the GET service

We'll never pass an error in now because we'll never call this in the POST handler.

* Update view to reference error

Also bring it inline with no returns required and start date.

* Update controller to use new service

* Update reference in validator

We changed the control to match the page and everything else so we need to update the validator to use the new reference.

* Fix the tests

* Remove unused static lookup

* Make the validator more robust

As our unit test confirmed it was checking for a value but would accept any value. Now it will only accept certain values.

* Update no-rtns-req validator

Now it is consistent with what we just did in the `ReasonValidator`.

* Add missing JSDoc comments

* Correct test title

* Add test for new submit service

---------

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>